Phase 6: split formatter.py into dates, numbers and textcase
formatter.py held four unrelated concerns in 523 lines: date parsing, numeric
coercion, money formatting, and a 340-line casing subsystem with its own YAML
data file and caching. The name described none of them. It is now three modules,
and the tests mirror them.

  utils/dates.py     51 lines   to_date
  utils/numbers.py   95 lines   to_decimal, to_float, to_int, round_to_millions
  utils/textcase.py  374 lines  TextFormatter, titlecase_name

No compatibility shim. Every symbol was internal, absent from __all__ and the
README, and all 27 importers are in-repo and updated here, so a shim would guard a
contract that does not exist. Verified by AST symbol diff that all 11 top-level
symbols moved exactly once and byte-identical.

round_to_millions stays with the coercers rather than moving to textcase or
earning its own module: its two call sites are display code that never touches
casing, and putting a money formatter inside the unit a NASA plugin would lift
makes that boundary worse, not better.

Removed current_fiscal_year, which had no callers anywhere, no test, and no
export. The endpoints whose docstrings mention defaulting to the current fiscal
year are defaulted server-side. Worth owning that I had "fixed" its double clock
read earlier in this branch, so that was a fix to dead code, and the changelog
entry for it is withdrawn.

Wired up PAREN_UPPERCASE_MAX_LEN, which existed only as a definition while
to_sentence_case hardcoded the same 9 as its default, and collapsed an
unreachable else in the parenthetical handler: the inverse condition returns
above it and neither operand changes in between. Ten parenthetical inputs give
identical output.

Tests: test_formatter.py became test_dates.py and test_numbers.py, its titlecase
class merged into test_textcase.py (renamed from test_text_formatter.py), and
three near-identical autouse fixtures plus a hand-rolled variant collapsed onto
one casing_from context manager over a shared list. Nine function-local imports
hoisted. Checked for fixture interference in both orders and per class, since the
merged block patches open while its neighbours manipulate the cache: safe,
because the lookups invalidate on list identity, which its own test pins.

The golden suite takes the head of a two-row page rather than calling first(),
after the recipient category was seen answering limit=1 with zero rows and then,
on re-probing, stopped doing so. Treated as upstream flakiness: I had written
characterization tests pinning it as a defect and removed them when it proved
unreproducible.
