+++ src/usaspending/models/award.py
@@ -9,7 +9,9 @@ from typing import TYPE_CHECKING, Any, ClassVar
 
 from ..exceptions import ValidationError
 from ..logging_config import USASpendingLogger
-from ..utils.formatter import TextFormatter, to_date, to_decimal, to_int
+from ..utils.dates import to_date
+from ..utils.numbers import to_decimal, to_int
+from ..utils.textcase import TextFormatter
 from .agency import Agency
 from .award_identifier import parse_award_identifier
 from .award_types import DOWNLOAD_TYPES
