+++ src/usaspending/models/grant.py
@@ -6,7 +6,7 @@ from decimal import Decimal
 from functools import cached_property
 from typing import TYPE_CHECKING, Any, ClassVar
 
-from ..utils.formatter import to_decimal
+from ..utils.numbers import to_decimal
 from .award import Award
 
 if TYPE_CHECKING:
