+++ tests/models/test_subtier_agency.py
@@ -2,7 +2,7 @@
 
 from tests.utils import assert_decimal_equal
 from usaspending.models.subtier_agency import SubTierAgency
-from usaspending.utils.formatter import titlecase_name
+from usaspending.utils.textcase import titlecase_name
 
 
 class TestSubTierAgency:
